requirements.txt
asgiref==3.8.1
Django==5.1.4
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
libmagic==1.0
pillow==11.1.0
PyJWT==2.10.1
python-magic==0.4.27
sqlparse==0.5.3
tzdata==2024.2
django-cors-headers
